move all parameters utils to parameters/utils files
