+++ frontend/src/metabase/dashboard/actions.js
@@ -17,12 +17,15 @@ import {
   setParameterDefaultValue as setParamDefaultValue,
   getMappingsByParameter,
   getDashboardParametersWithFieldMetadata,
-} from "metabase/meta/Dashboard";
-import { applyParameters, questionUrlWithParameters } from "metabase/meta/Card";
+} from "metabase/parameters/utils/dashboards";
+import {
+  applyParameters,
+  questionUrlWithParameters,
+} from "metabase/parameters/utils/cards";
 import {
   getParameterValuesBySlug,
   getParameterValuesByIdFromQueryParams,
-} from "metabase/meta/Parameter";
+} from "metabase/parameters/utils/parameter-values";
 import * as Urls from "metabase/lib/urls";
 import { SIDEBAR_NAME } from "metabase/dashboard/constants";
