+++ frontend/src/metabase/parameters/components/ParameterValueWidget.jsx
@@ -23,11 +23,9 @@ import {
   makeGetMergedParameterFieldValues,
 } from "metabase/selectors/metadata";
 
-import {
-  getParameterIconName,
-  deriveFieldOperatorFromParameter,
-} from "metabase/meta/Parameter";
-import { isDashboardParameterWithoutMapping } from "metabase/meta/Dashboard";
+import { getParameterIconName } from "metabase/parameters/utils/ui";
+import { deriveFieldOperatorFromParameter } from "metabase/parameters/utils/operators";
+import { isDashboardParameterWithoutMapping } from "metabase/parameters/utils/dashboards";
 
 import S from "./ParameterWidget.css";
