+++ frontend/src/metabase/parameters/utils/filters.unit.spec.js
@@ -1,4 +1,7 @@
-import { dimensionFilterForParameter } from "./field-filters";
+import {
+  dimensionFilterForParameter,
+  getTagOperatorFilterForParameter,
+} from "./filters";
 
 describe("parameters/utils/field-filters", () => {
   describe("dimensionFilterForParameter", () => {
@@ -119,6 +122,26 @@ describe("parameters/utils/field-filters", () => {
 
       const predicate = dimensionFilterForParameter({ type: "category" });
       expect(predicate(locationDimension)).toBe(false);
+    });
+  });
+
+  describe("getTagOperatorFilterForParameter", () => {
+    it("should return a predicate that evaluates to true for a template tag that has the same subtype operator as the given parameter", () => {
+      const predicate = getTagOperatorFilterForParameter({
+        type: "string/starts-with",
+      });
+      const templateTag1 = {
+        "widget-type": "string/starts-with",
+      };
+      const templateTag2 = {
+        "widget-type": "foo/starts-with",
+      };
+      const templateTag3 = {
+        "widget-type": "string/ends-with",
+      };
+      expect(predicate(templateTag1)).toBe(true);
+      expect(predicate(templateTag2)).toBe(true);
+      expect(predicate(templateTag3)).toBe(false);
     });
   });
 });
