+++ frontend/src/metabase/parameters/utils/fixtures/dashboard-with-boolean-parameter-fixture.json
