+++ frontend/src/metabase/parameters/utils/template-tag-options.js
@@ -1,8 +1,6 @@
-import {
-  areFieldFilterOperatorsEnabled,
-  getOperatorDisplayName,
-} from "./internal";
-import { fieldFilterForParameter } from "./field-filters";
+import { areFieldFilterOperatorsEnabled } from "./parameter-type";
+import { getOperatorDisplayName } from "./operators";
+import { fieldFilterForParameter } from "./filters";
 
 import {
   OPTIONS_WITH_OPERATOR_SUBTYPES,
