+++ frontend/src/metabase/public/containers/PublicQuestion.jsx
@@ -16,12 +16,12 @@ import type { ParameterValues } from "metabase-types/types/Parameter";
 import {
   getParameterValuesBySlug,
   getParameterValuesByIdFromQueryParams,
-} from "metabase/meta/Parameter";
+} from "metabase/parameters/utils/parameter-values";
 import {
   getParametersFromCard,
   getValueAndFieldIdPopulatedParametersFromCard,
   applyParameters,
-} from "metabase/meta/Card";
+} from "metabase/parameters/utils/cards";
 
 import {
   PublicApi,
